requirements.txt
./streamlit-1.32.2-py2.py3-none-any.whl
openai
langchain
langchain_openai
langchainhub
streamlit-audiorecorder
google-auth
streamlit-oauth
pytest
isort
streamlit-authenticator
google-api-python-client
google-auth-httplib2
google-auth-oauthlib